Extract interface from VersionExtractor
